requirements.txt
google-genai
requests
pandas
ta
streamlit
python-dotenv
yfinance
